factor out ImportMap::load
